requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
chromadb
langchain
langchain-community
langchain-text-splitters
langchain-chroma
rank_bm25
sentence-transformers
httpx
structlog
tenacity
pypdf
pdfplumber
pillow
ragas
datasets
streamlit
plotly